Conflicts:
	Library/Formula/libffi.rb
	Library/Formula/qt.rb
	Library/Homebrew/tap_migrations.rb

+++ Library/Formula/qt.rb
@@ -23,7 +23,7 @@ class Qt < Formula
     sha1 "668ac1a65811e0ff23230a698725b383c61c1d13" => :mountain_lion
   end
 
-  head "git://gitorious.org/qt/qt.git", :branch => '4.8'
+  head "https://gitorious.org/qt/qt.git", :branch => "4.8"
 
   option :universal
   option 'with-qt3support', 'Build with deprecated Qt3Support module support'
@@ -101,10 +101,8 @@ class Qt < Formula
     system "make install"
 
     # what are these anyway?
-    if OS.mac?
-      (bin+'pixeltool.app').rmtree
-      (bin+'qhelpconverter.app').rmtree
-    end
+    (bin+'pixeltool.app').rmtree if OS.mac?
+    (bin+'qhelpconverter.app').rmtree if OS.mac?
     # remove porting file for non-humans
     (prefix+'q3porting.xml').unlink if build.without? 'qt3support'
 
@@ -118,7 +116,7 @@ class Qt < Formula
       include.install_symlink path => path.parent.basename(".framework")
     end
 
-    Pathname.glob("#{bin}/*.app") { |app| mv app, prefix } if OS.mac?
+    Pathname.glob("#{bin}/*.app") { |app| mv app, prefix }
   end
 
   test do
